Crate frame

Made for

FRAME

  ______   ______    ________   ___ __ __   ______
 /_____/\ /_____/\  /_______/\ /__//_//_/\ /_____/\
 \::::_\/_\:::_ \ \ \::: _  \ \\::\| \| \ \\::::_\/_
  \:\/___/\\:(_) ) )_\::(_)  \ \\:.      \ \\:\/___/\
   \:::._\/ \: __ `\ \\:: __  \ \\:.\-/\  \ \\::___\/_
    \:\ \    \ \ `\ \ \\:.\ \  \ \\. \  \  \ \\:\____/\
     \_\/     \_\/ \_\/ \__\/\__\/ \__\/ \__\/ \_____\/

Framework for Runtime Aggregation of Modularized Entities: Substrate’s State Transition Function (Runtime) Framework.

Warning: Experimental

This crate and all of its content is experimental, and should not yet be used in production.

Getting Started

TODO: link to developer_hub::polkadot_sdk::frame. The developer_hub hasn’t been published yet, this can be updated once it is linkable.

Re-exports

pub use log;

Modules

arithmetic

The arithmetic types used for safe math.

deps

Access to all of the dependencies of this crate. In case the re-exports are not enough, this module can be used.

derive

All derive macros used in frame.

prelude

The main prelude of FRAME.

primitives

Low level primitive types used in FRAME pallets.

runtime

All of the types and tools needed to build FRAME-based runtimes.

testing_prelude

The main testing prelude of FRAME.

traits

All traits often used in FRAME pallets.

Attribute Macros

pallet

Exports the main pallet macro. This can wrap a mod pallet and will transform it into being a pallet, eg #[frame::pallet] mod pallet { .. }.