Struct frame::arithmetic::rational::Rational128

pub struct Rational128(_, _);

A wrapper for any rational number with a 128 bit numerator and denominator.

Implementations

impl Rational128

pub fn zero() -> Rational128

Zero.

pub fn one() -> Rational128

One

pub fn is_zero(&self) -> bool

If it is zero or not

pub fn from(n: u128, d: u128) -> Rational128

Build from a raw n/d.

pub fn from_unchecked(n: u128, d: u128) -> Rational128

Build from a raw n/d. This could lead to / 0 if not properly handled.

pub fn n(&self) -> u128

Return the numerator.

pub fn d(&self) -> u128

Return the denominator.

pub fn to_den(self, den: u128) -> Option<Rational128>

Convert self to a similar rational number where denominator is the given den. This only returns if the result is accurate. None is returned if the result cannot be accurately calculated.

pub fn lcm(&self, other: &Rational128) -> Option<u128>

Get the least common divisor of self and other.

This only returns if the result is accurate. None is returned if the result cannot be accurately calculated.

pub fn lazy_saturating_add(self, other: Rational128) -> Rational128

A saturating add that assumes self and other have the same denominator.

pub fn lazy_saturating_sub(self, other: Rational128) -> Rational128

A saturating subtraction that assumes self and other have the same denominator.

pub fn checked_add( self, other: Rational128 ) -> Result<Rational128, &'static str>

Addition. Simply tries to unify the denominators and add the numerators.

Overflow might happen during any of the steps. Error is returned in such cases.

pub fn checked_sub( self, other: Rational128 ) -> Result<Rational128, &'static str>

Subtraction. Simply tries to unify the denominators and subtract the numerators.

Overflow might happen during any of the steps. None is returned in such cases.