Trait frame::arithmetic::Normalizable

pub trait Normalizable<T> {
    // Required method
    fn normalize(&self, targeted_sum: T) -> Result<Vec<T, Global>, &'static str>;
}

A collection-like object that is made of values of type T and can normalize its individual values around a centric point.

Note that the order of items in the collection may affect the result.

Required Methods

fn normalize(&self, targeted_sum: T) -> Result<Vec<T, Global>, &'static str>

Normalize self around targeted_sum.

Only returns Ok if the new sum of results is guaranteed to be equal to targeted_sum. Else, returns an error explaining why it failed to do so.