Module frame::deps

Access to all of the dependencies of this crate. In case the re-exports are not enough, this module can be used.

Any time one uses this module to access a dependency, you can have a moment to think about whether this item could have been placed in any of the other modules and preludes in this crate. In most cases, hopefully the answer is yes.

Re-exports

pub use frame_support;

pub use frame_system;

pub use sp_arithmetic;

pub use sp_core;

pub use sp_io;

pub use sp_runtime;

pub use sp_std;

pub use parity_scale_codec as codec;

pub use scale_info;

pub use frame_executive;

pub use sp_api;

pub use sp_block_builder;

pub use sp_consensus_aura;

pub use sp_consensus_grandpa;

pub use sp_inherents;

pub use sp_offchain;

pub use sp_version;