Module frame::runtime::apis

Types and traits for runtimes that implement runtime APIs.

A testing runtime should not need this.

A non-testing runtime should have this enabled, as such:

use frame::runtime::{prelude::*, apis::{*,}};

Re-exports

pub use sp_api;

Modules

digests

Aura (Authority-Round) digests

ed25519

inherents

sr25519

Macros

_std_disabled

Enable/disable the given code depending on not(feature = "std") being enabled for the crate or not.

_std_enabled

Enable/disable the given code depending on feature = "std" being enabled for the crate or not.

create_apis_vec

Create a vector of Api declarations.

decl_runtime_apis

impl_runtime_apis

Macro to implement runtime APIs.

mock_impl_runtime_apis

std_disabled

Enable/disable the given code depending on not(feature = "std") being enabled for the crate or not.

std_enabled

Enable/disable the given code depending on feature = "std" being enabled for the crate or not.

vec

Creates a Vec containing the arguments.

Structs

ApiRef

Auxiliary wrapper that holds an api instance and binds it to the given lifetime.

CallApiAtParams

Parameters for CallApiAt::call_api_at.

CheckInherentsResult

The result of checking inherents.

EquivocationProof

Proof of voter misbehavior on a given set id. Misbehavior/equivocation in GRANDPA happens when a voter votes on the same round (either at prevote or precommit stage) for different blocks. Proving is achieved by collecting the signed messages of conflicting votes.

Extensions

Stores extensions that should be made available through the externalities.

GrandpaJustification

A GRANDPA justification for block finality, it includes a commit message and an ancestry proof including all headers routing all precommit target blocks to the commit target block. Due to the current voting strategy the precommit targets should be the same as the commit target, since honest voters don’t vote past authority set change blocks.

InherentData

Inherent data to include in a block.

KeyTypeId

An identifier for a type of cryptographic key.

OpaqueKeyOwnershipProof

An opaque type used to represent the key ownership proof at the runtime API boundary. The inner value is an encoded representation of the actual key ownership proof which will be parameterized when defining the runtime. At the runtime API boundary this type is unknown and as such we keep this opaque representation, implementors of the runtime API will have to make sure that all usages of OpaqueKeyOwnershipProof refer to the same type.

OpaqueMetadata

Stores the encoded RuntimeMetadata for the native side as opaque type.

OverlayedChanges

The set of changes that are overlaid onto the backend.

RuntimeVersion

Runtime version. This should not be thought of as classic Semver (major/minor/tiny). This triplet have different semantics and mis-interpretation could cause problems. In particular: bug fixes should result in an increment of spec_version and possibly authoring_version, absolutely not impl_version since they change the semantics of the runtime.

ScheduledChange

A scheduled change of authority set.

Slot

Unit type wrapper that represents a slot.

SlotDuration

A slot duration defined in milliseconds.

StorageProof

A proof that some set of key-value pairs are included in the storage trie. The proof contains the storage values so that the partial storage backend can be reconstructed by a verifier that does not already have access to the key-value pairs.

TrieBackend

Patricia trie-based backend. Transaction type is an overlay of changes to commit.

TrieBackendBuilder

Builder for creating a TrieBackend.

VersionedAuthorityList

An AuthorityList that is encoded with a version specifier. The encoding version is updated any time the AuthorityList type changes. This ensures that encodings of different versions of an AuthorityList are differentiable. Attempting to decode an authority list with an unknown version will fail.

Enums

ApiError

An error describing which API call failed.

BlockId

Something to identify a block.

CallContext

The context in which a call is done.

Equivocation

Wrapper object for GRANDPA equivocation proofs, useful for unifying prevote and precommit equivocations under a common type.

RuntimeString

A string that wraps a &'static str in the runtime and String/Vec<u8> on decode.

StateVersion

Different possible state version.

TransactionOutcome

Describes on what should happen with a storage transaction.

Constants

AURA_ENGINE_ID

The ConsensusEngineId of AuRa.

AUTHORITIES_CALL

WASM function call to get current GRANDPA authorities.

CLIENT_LOG_TARGET

The log target to be used by client code.

GRANDPA_AUTHORITIES_KEY

The storage key for the current set of weighted Grandpa authorities. The value stored is an encoded VersionedAuthorityList.

GRANDPA_ENGINE_ID

The ConsensusEngineId of GRANDPA.

KEY_TYPE

Key type for GRANDPA module.

MAX_EXTRINSIC_DEPTH

Maximum nesting level for extrinsics.

PENDING_CHANGE_CALL

WASM function call to check for pending changes.

RUNTIME_API_INFO_SIZE

The number of bytes required to encode a RuntimeApiInfo.

RUNTIME_LOG_TARGET

The log target to be used by runtime code.

STORAGE_PREFIX

Persistent storage prefix used by the Offchain Worker API when creating a DB key.

Traits

AccountNonceApi

The API to query account nonce.

ApiExt

Extends the runtime api implementation with some common functionality.

AsTrieBackend

Something that can be converted into a TrieBackend.

AuraApi

API necessary for block authorship with aura.

BlockBuilder

The BlockBuilder api trait that provides the required functionality for building a block.

BlockT

Something which fulfills the abstract idea of a Substrate block. It has types for Extrinsic pieces of information as well as a Header.

CallApiAt

Something that can call into the an api at a given block.

ConstructRuntimeApi

Something that can be constructed to a runtime api.

Core

The Core runtime api that every Substrate runtime needs to implement.

Decode

Trait that allows zero-copy read of value-references from slices in LE format.

DecodeLimit

Extension trait to Decode for decoding with a maximum recursion depth.

Encode

Trait that allows zero-copy write of value-references to slices in LE format.

Extension

Marker trait for types that should be registered as Externalities extension.

GrandpaApi

APIs for integrating the GRANDPA finality gadget into runtimes. This should be implemented on the runtime side.

HashT

Abstraction around hashing

Hasher

Trait describing an object that can hash a slice of bytes. Used to abstract other types over the hashing algorithm. Defines a single hash method and an Out associated type with the necessary bounds.

HeaderT

Something which fulfills the abstract idea of a Substrate header. It has types for a Number, a Hash and a Hashing. It provides access to an extrinsics_root, state_root and parent_hash, as well as a digest and a block number.

Metadata

The Metadata api trait that returns metadata for the runtime.

OffchainWorkerApi

The offchain worker api.

ProvideRuntimeApi

Something that provides a runtime api.

RuntimeApiInfo

Something that provides information about a runtime api.

SessionKeys

Session keys runtime api.

StateBackend

A state backend is used to read state data and can have changes committed to it.

TaggedTransactionQueue

The TaggedTransactionQueue api trait for interfering with the transaction queue.

Functions

check_equivocation_proof

Verifies the equivocation proof by making sure that both votes target different blocks and that its signatures are valid.

check_message_signature

Check a message signature by encoding the message as a localized payload and verifying the provided signature using the expected authority id.

check_message_signature_with_buffer

Check a message signature by encoding the message as a localized payload and verifying the provided signature using the expected authority id. The encoding necessary to verify the signature will be done using the given buffer, the original content of the buffer will be cleared.

deserialize_runtime_api_info

Deserialize the runtime API info serialized by serialize_runtime_api_info.

init_runtime_logger

Init the RuntimeLogger.

localized_payload

Encode round message localized to a given round and set id.

localized_payload_with_buffer

Encode round message localized to a given round and set id using the given buffer. The given buffer will be cleared and the resulting encoded payload will always be written to the start of the buffer.

serialize_runtime_api_info

Crude and simple way to serialize the RuntimeApiInfo into a bunch of bytes.

sign_message

Localizes the message to the given set and round and signs the payload.

Type Definitions

ApiId

The identity of a particular API interface that the runtime might provide.

ApisVec

A vector of pairs of ApiId and a u32 for version.

ApplyExtrinsicResult

The result of applying of an extrinsic.

AuthorityId

Identity of a Grandpa authority.

AuthorityList

A list of Grandpa authorities with associated weights.

AuthorityPair

The grandpa crypto scheme defined via the keypair type.

AuthoritySignature

Signature for a Grandpa authority.

AuthorityWeight

The weight of an authority.

CatchUp

A catch up message for this chain’s block type.

Commit

A commit message for this chain’s block type.

CompactCommit

A compact commit message for this chain’s block type.

HashingFor

Extract the hashing type for a block.

InMemoryBackend

Trie backend with in-memory storage.

Message

A GRANDPA message for a substrate chain.

NumberFor

Extract the number type for a block.

Precommit

A precommit message for this chain’s block type.

Prevote

A prevote message for this chain’s block type.

PrimaryPropose

A primary propose message for this chain’s block type.

ProofRecorder

A type that records all accessed trie nodes and generates a proof out of it.

RoundNumber

The round indicator.

SetId

The monotonic identifier of a GRANDPA set of authorities.

SignedMessage

A signed message.

StorageChanges

TransactionValidity

Information on a transaction’s validity and, if valid, on how it relates to other transactions.

Derive Macros

Decode

Derive parity_scale_codec::Decode and for struct and enum.

Encode

Derive parity_scale_codec::Encode and parity_scale_codec::EncodeLike for struct and enum.