Type Alias frame::runtime::apis::InMemoryBackend

pub type InMemoryBackend<H> = TrieBackend<MemoryDB<H, PrefixedKey<H>, Vec<u8, Global>>, H, LocalTrieCache<H>>;

Trie backend with in-memory storage.

Aliased Type

struct InMemoryBackend<H> { /* private fields */ }

Implementations

impl<H> TrieBackend<MemoryDB<H, PrefixedKey<H>, Vec<u8, Global>>, H, LocalTrieCache<H>>where H: Hasher, <H as Hasher>::Out: Codec + Ord,

pub fn update<T>( &self, changes: T, state_version: StateVersion ) -> TrieBackend<MemoryDB<H, PrefixedKey<H>, Vec<u8, Global>>, H, LocalTrieCache<H>>where T: IntoIterator<Item = (Option<ChildInfo>, Vec<(Vec<u8, Global>, Option<Vec<u8, Global>>), Global>)>,

Copy the state, with applied updates

pub fn insert<T>(&mut self, changes: T, state_version: StateVersion)where T: IntoIterator<Item = (Option<ChildInfo>, Vec<(Vec<u8, Global>, Option<Vec<u8, Global>>), Global>)>,

Insert values into backend trie.

pub fn update_backend( &self, root: <H as Hasher>::Out, changes: MemoryDB<H, PrefixedKey<H>, Vec<u8, Global>> ) -> TrieBackend<MemoryDB<H, PrefixedKey<H>, Vec<u8, Global>>, H, LocalTrieCache<H>>

Merge trie nodes into this backend.

pub fn apply_transaction( &mut self, root: <H as Hasher>::Out, transaction: MemoryDB<H, PrefixedKey<H>, Vec<u8, Global>> )

Apply the given transaction to this backend and set the root to the given value.

pub fn eq( &self, other: &TrieBackend<MemoryDB<H, PrefixedKey<H>, Vec<u8, Global>>, H, LocalTrieCache<H>> ) -> bool

Compare with another in-memory backend.

impl<S, H, C> TrieBackend<S, H, C>where S: TrieBackendStorage<H>, H: Hasher, C: TrieCacheProvider<H> + Send + Sync, <H as Hasher>::Out: Codec,

pub fn essence(&self) -> &TrieBackendEssence<S, H, C>

Get backend essence reference.

pub fn backend_storage_mut(&mut self) -> &mut S

Get backend storage reference.

pub fn backend_storage(&self) -> &S

Get backend storage reference.

pub fn set_root(&mut self, root: <H as Hasher>::Out)

Set trie root.

pub fn root(&self) -> &<H as Hasher>::Out

Get trie root.

pub fn into_storage(self) -> S

Consumes self and returns underlying storage.

pub fn extract_proof(self) -> Option<StorageProof>

Extract the StorageProof.

This only returns Some when there was a recorder set.