Type Alias frame::runtime::apis::StorageChanges

pub type StorageChanges<Block> = StorageChanges<<<Block as Block>::Header as Header>::Hashing>;

Aliased Type

struct StorageChanges<Block> {
    pub main_storage_changes: Vec<(Vec<u8, Global>, Option<Vec<u8, Global>>), Global>,
    pub child_storage_changes: Vec<(Vec<u8, Global>, Vec<(Vec<u8, Global>, Option<Vec<u8, Global>>), Global>), Global>,
    pub offchain_storage_changes: Vec<((Vec<u8, Global>, Vec<u8, Global>), OffchainOverlayedChange), Global>,
    pub transaction: MemoryDB<<<Block as Block>::Header as Header>::Hashing, PrefixedKey<<<Block as Block>::Header as Header>::Hashing>, Vec<u8, Global>>,
    pub transaction_storage_root: <<<Block as Block>::Header as Header>::Hashing as Hasher>::Out,
    pub transaction_index_changes: Vec<IndexOperation, Global>,
}

Fields
main_storage_changes: Vec<(Vec<u8, Global>, Option<Vec<u8, Global>>), Global>

All changes to the main storage.

A value of None means that it was deleted.

child_storage_changes: Vec<(Vec<u8, Global>, Vec<(Vec<u8, Global>, Option<Vec<u8, Global>>), Global>), Global>

All changes to the child storages.

offchain_storage_changes: Vec<((Vec<u8, Global>, Vec<u8, Global>), OffchainOverlayedChange), Global>

Offchain state changes to write to the offchain database.

transaction: MemoryDB<<<Block as Block>::Header as Header>::Hashing, PrefixedKey<<<Block as Block>::Header as Header>::Hashing>, Vec<u8, Global>>

A transaction for the backend that contains all changes from main_storage_changes and from child_storage_changes. offchain_storage_changes.

transaction_storage_root: <<<Block as Block>::Header as Header>::Hashing as Hasher>::Out

The storage root after applying the transaction.

transaction_index_changes: Vec<IndexOperation, Global>

Changes to the transaction index,