Struct frame::runtime::prelude::Executive

pub struct Executive<System, Block, Context, UnsignedValidator, AllPalletsWithSystem, OnRuntimeUpgrade = ()>(_);

All of the types related to the FRAME runtime executive. Main entry point for certain runtime actions as e.g. execute_block.

Generic parameters:

System: Something that implements frame_system::Config

Block: The block type of the runtime

Context: The context that is used when checking an extrinsic.

UnsignedValidator: The unsigned transaction validator of the runtime.

AllPalletsWithSystem: Tuple that contains all pallets including frame system pallet. Will be used to call hooks e.g. on_initialize.

OnRuntimeUpgrade: Custom logic that should be called after a runtime upgrade. Modules are already called by AllPalletsWithSystem. It will be called before all modules will be called.

Implementations

impl<System, Block, Context, UnsignedValidator, AllPalletsWithSystem, COnRuntimeUpgrade> Executive<System, Block, Context, UnsignedValidator, AllPalletsWithSystem, COnRuntimeUpgrade>where System: Config + EnsureInherentsAreFirst<Block>, Block: Block<Header = <<System as Config>::Block as HeaderProvider>::HeaderT, Hash = <System as Config>::Hash>, Context: Default, AllPalletsWithSystem: OnRuntimeUpgrade + OnInitialize<<<<System as Config>::Block as HeaderProvider>::HeaderT as Header>::Number> + OnIdle<<<<System as Config>::Block as HeaderProvider>::HeaderT as Header>::Number> + OnFinalize<<<<System as Config>::Block as HeaderProvider>::HeaderT as Header>::Number> + OffchainWorker<<<<System as Config>::Block as HeaderProvider>::HeaderT as Header>::Number>, COnRuntimeUpgrade: OnRuntimeUpgrade, <Block as Block>::Extrinsic: Checkable<Context> + Codec, <<Block as Block>::Extrinsic as Checkable<Context>>::Checked: Applyable + GetDispatchInfo, <<<Block as Block>::Extrinsic as Checkable<Context>>::Checked as Applyable>::Call: Dispatchable<Info = DispatchInfo, PostInfo = PostDispatchInfo>, <<<<Block as Block>::Extrinsic as Checkable<Context>>::Checked as Applyable>::Call as Dispatchable>::RuntimeOrigin: From<Option<<System as Config>::AccountId>>, UnsignedValidator: ValidateUnsigned<Call = <<<Block as Block>::Extrinsic as Checkable<Context>>::Checked as Applyable>::Call>,

pub fn execute_on_runtime_upgrade() -> Weight

Execute all OnRuntimeUpgrade of this runtime, and return the aggregate weight.

pub fn initialize_block( header: &<<System as Config>::Block as HeaderProvider>::HeaderT )

Start the execution of a particular block.

pub fn execute_block(block: Block)

Actually execute all transitions for block.

pub fn finalize_block( ) -> <<System as Config>::Block as HeaderProvider>::HeaderT

Finalize the block - it is up the caller to ensure that all header fields are valid except state-root.

pub fn apply_extrinsic( uxt: <Block as Block>::Extrinsic ) -> Result<Result<(), DispatchError>, TransactionValidityError>

Apply extrinsic outside of the block execution function.

This doesn’t attempt to validate anything regarding the block, but it builds a list of uxt hashes.

pub fn validate_transaction( source: TransactionSource, uxt: <Block as Block>::Extrinsic, block_hash: <Block as Block>::Hash ) -> Result<ValidTransaction, TransactionValidityError>

Check a given signed transaction for validity. This doesn’t execute any side-effects; it merely checks whether the transaction would panic if it were included or not.

Changes made to storage should be discarded.

pub fn offchain_worker( header: &<<System as Config>::Block as HeaderProvider>::HeaderT )

Start an offchain worker and generate extrinsics.