Module frame::traits

All traits often used in FRAME pallets.

Note that types implementing these traits can also be found in this module.

Modules

defensive_prelude

Prelude module for all defensive traits to be imported at once.

schedule

Traits and associated utilities for scheduling dispatchables in FRAME.

tokens

Traits for working with tokens and their associated datastructures.

Macros

parameter_types

Create new implementations of the Get trait.

Structs

AccountIdLookup

A lookup implementation returning the AccountId from a MultiAddress.

ActiveIssuanceOf

A non-const Get implementation parameterised by a Currency impl which provides the result of active_issuance.

AppendZerosInput

Input that adds infinite number of zero after wrapped input.

AsContains

Adapter struct for turning an OrderedMembership impl into a Contains impl.

AsEnsureOriginWithArg

Backing

Some amount of backing from a group. The precise definition of what it means to “back” something is left flexible.

BadOrigin

An error type that indicates that the origin is invalid.

BlakeTwo256

Blake2-256 Hash implementation.

CallMetadata

The function and pallet name of the Call.

CheckedReduceBy

A TryMorph implementation to reduce a scalar by a particular amount, checking for underflow.

ClassCountOf

ClearFilterGuard

Guard type for clearing all pushed constraints from a FilterStack and reinstating them when dropped.

ConstBool

Const getter for a basic type.

ConstI8

Const getter for a basic type.

ConstI16

Const getter for a basic type.

ConstI32

Const getter for a basic type.

ConstI64

Const getter for a basic type.

ConstI128

Const getter for a basic type.

ConstU8

Const getter for a basic type.

ConstU16

Const getter for a basic type.

ConstU32

Const getter for a basic type.

ConstU64

Const getter for a basic type.

ConstU128

Const getter for a basic type.

ConvertInto

A structure that performs standard conversion using the standard Rust conversion traits.

ConvertToValue

Adapter which turns a Get implementation into a Convert implementation which always returns in the same value no matter the input.

CrateVersion

The version of a crate.

EitherOf

“OR gate” implementation of EnsureOrigin, Success type for both L and R must be equal.

EitherOfDiverse

“OR gate” implementation of EnsureOrigin allowing for different Success types for L and R, with them combined using an Either type.

EnqueueWithOrigin

Adapter type to transform an EnqueueMessage with an origin into a HandleMessage impl.

EnsureOriginEqualOrHigherPrivilege

EnsureOrigin implementation that checks that an origin has equal or higher privilege compared to the expected Origin.

EqualPrivilegeOnly

Implementation of PrivilegeCmp that only checks for equal origins.

Equals

An implementation of Contains which contains only equal members to T.

EverythingBut

A Contains implementation that contains everything except the values in Exclude.

FilterStackGuard

Guard type for pushing a constraint to a FilterStack and popping when dropped.

Footprint

The resource footprint of a bunch of blobs. We assume only the number of blobs and their total size in bytes matter.

FromContainsPair

Converter struct to use a ContainsPair implementation for a Contains bound.

GetDefault

Implement Get by returning Default for any type that implements Default.

Identity

A structure that performs identity conversion.

IdentityLookup

A lookup implementation returning the input value.

InsideBoth

A Contains implementation which contains all members of These which are also members of Those.

IsInVec

Trivial utility for implementing Contains/OrderedMembership with a Vec.

Keccak256

Keccak-256 Hash implementation.

LinearStoragePrice

A storage price that increases linearly with the number of elements and their size.

LookupError

An error that indicates that a lookup failed.

MapSuccess

A derivative EnsureOrigin implementation. It mutates the Success result of an Original implementation with a given Mutator.

MorphInto

Implementation of Morph which converts between types using Into.

MorphWithUpperLimit

A TryMorph implementation to enforce an upper limit for a result of the outer morphed type.

NeverEnsureOrigin

EnsureOrigin implementation that always fails.

NoStorageVersionSet

Special marker struct if no storage version is set for a pallet.

NoopServiceQueues

Services queues by doing nothing.

PalletInfoData

Information regarding an instance of a pallet.

PhantomData

Zero-sized type used to mark things that “act like” they own a T.

ReduceBy

Mutator which reduces a scalar by a particular amount.

Replace

Morpher to disregard the source value and replace with another.

SignedExtensionMetadata

Information about a SignedExtension for the runtime metadata.

StorageInfo

Metadata about storage from the runtime.

StorageMapShim

A shim for placing around a storage item in order to use it as a StoredValue. Ideally this wouldn’t be needed as StorageValues should blanket implement StoredValues, however this would break the ability to have custom impls of StoredValue. The other workaround is to implement it directly in the macro.

StorageVersion

The storage version of a pallet.

TakeFirst

Implementation of Morph to retrieve just the first element of a tuple.

TheseExcept

A Contains implementation that contains all members of These excepting any members in Except.

TotalIssuanceOf

A non-const Get implementation parameterised by a Currency impl which provides the result of total_issuance.

TrackedStorageKey

Storage key with read/write tracking information.

TrailingZeroInput

Input that adds infinite number of zero after wrapped input.

TransformOrigin

Transform the origin of an EnqueueMessage via C::convert.

TryConvertInto

A structure that performs standard conversion using the standard Rust conversion traits.

TryMapSuccess

A derivative EnsureOrigin implementation. It mutates the Success result of an Original implementation with a given Mutator, allowing the possibility of an error to be returned from the mutator.

TryMorphInto

Implementation of TryMorph which attmepts to convert between types using TryInto.

TryWithMorphedArg

WithdrawReasons

Reasons for moving funds out of an account.

WrapperKeepOpaque

A wrapper for any type T which implement encode/decode in a way compatible with Vec<u8>.

WrapperOpaque

A wrapper for any type T which implement encode/decode in a way compatible with Vec<u8>.

Enums

BalanceStatus

Status of funds.

Everything

A Contains implementation that contains every value.

ExecuteOverweightError

Errors that can happen when attempting to execute an overweight message with ServiceQueues::execute_overweight().

ExistenceRequirement

Simple boolean for whether an account needs to be kept in existence.

Nothing

A Contains implementation that contains no value.

PollStatus

ProcessMessageError

Errors that can happen when attempting to process a message with ProcessMessage::process_message().

SafeModeError

The error type for SafeMode.

SameOrOther

Return type used when we need to return one of two items, each of the opposite direction or sign, with one (Same) being of the same type as the self or primary argument of the function that returned it.

SignedImbalance

Either a positive or a negative imbalance.

TransactionPauseError

The error type for TransactionPause.

Constants

STORAGE_VERSION_STORAGE_KEY_POSTFIX

The storage key postfix that is used to store the StorageVersion per pallet.

Traits

AccountIdConversion

This type can be converted into and possibly from an AccountId (which itself is generic).

AccountTouch

Trait for creating an asset account with a deposit taken from a designated depositor specified by the client.

AppVerify

Means of signature verification of an application key.

Applyable

An “executable” piece of information, used by the standard Substrate Executive in order to enact a piece of extrinsic information by marshalling and dispatching to a named function call.

AtLeast32Bit

A meta trait for arithmetic.

AtLeast32BitUnsigned

A meta trait for arithmetic. Same as AtLeast32Bit, but also bounded to be unsigned.

BlindCheckable

A “checkable” piece of information, used by the standard Substrate Executive in order to check the validity of a piece of extrinsic information, usually by verifying the signature. Implement for pieces of information that don’t require additional context in order to be checked.

Block

Something which fulfills the abstract idea of a Substrate block. It has types for Extrinsic pieces of information as well as a Header.

BlockIdTo

Something that can convert a BlockId to a number or a hash.

BlockNumberProvider

Get current block number

BuildGenesisConfig

A trait to define the build function of a genesis config for both runtime and pallets.

CallerTrait

The trait implemented by the overarching enumeration of the different pallets’ origins. Unlike OriginTrait impls, this does not include any kind of dispatch/call filter. Also, this trait is more flexible in terms of how it can be used: it is a Parameter and Member, so it can be used as dispatchable parameters as well as in storage items.

ChangeMembers

Trait for type that can handle incremental changes to a set of account IDs.

CheckEqual

Something that can be checked for equality and printed out to a debug channel if bad.

Checkable

Extract the digest type for a block. A “checkable” piece of information, used by the standard Substrate Executive in order to check the validity of a piece of extrinsic information, usually by verifying the signature. Implement for pieces of information that require some additional context Context in order to be checked.

CheckedAdd

Performs addition that returns None instead of wrapping around on overflow.

CheckedConversion

Convenience type to work around the highly unergonomic syntax needed to invoke the functions of overloaded generic traits, in this case TryFrom and TryInto.

CheckedDiv

Performs division that returns None instead of panicking on division by zero and instead of wrapping around on underflow and overflow.

CheckedMul

Performs multiplication that returns None instead of wrapping around on underflow or overflow.

CheckedShl

Performs a left shift that returns None on shifts larger than or equal to the type width.

CheckedShr

Performs a right shift that returns None on shifts larger than or equal to the type width.

CheckedSub

Performs subtraction that returns None instead of wrapping around on underflow.

Clear

Trait for things that can be clear (have no bits set). For numeric types, essentially the same as Zero.

Consideration

Some sort of cost taken from account temporarily in order to offset the cost to the chain of holding some data Footprint in state.

Contains

A trait for querying whether a type can be said to “contain” a value.

ContainsLengthBound

A trait for querying bound for the length of an implementation of Contains

ContainsPair

A trait for querying whether a type can be said to “contain” a pair-value.

Convert

Infallible conversion trait. Generic over both source and destination types.

ConvertBack

Reversing infallible conversion trait. Generic over both source and destination types.

Currency

Abstraction over a fungible assets system.

Defensive

A trait to handle errors and options when you are really sure that a condition must hold, but not brave enough to expect on it, or a default fallback value makes more sense.

DefensiveMax

Defensively calculates the maximum of two values.

DefensiveMin

Defensively calculates the minimum of two values.

DefensiveOption

Subset of methods similar to Defensive that can only work for a Option.

DefensiveResult

Subset of methods similar to Defensive that can only work for a Result.

DefensiveSaturating

A variant of Defensive with the same rationale, for the arithmetic operations where in case an infallible operation fails, it saturates.

DefensiveTruncateFrom

Construct an object by defensively truncating an input if the TryFrom conversion fails.

DisabledValidators

Trait used to check whether a given validator is currently disabled and should not be participating in consensus (e.g. because they equivocated).

Dispatchable

A lazy call (module function and argument values) that can be executed via its dispatch method.

EnqueueMessage

Can enqueue messages for multiple origins.

Ensure

EnsureAdd

Performs addition that returns ArithmeticError instead of wrapping around on overflow.

EnsureAddAssign

Performs self addition that returns ArithmeticError instead of wrapping around on overflow.

EnsureDiv

Performs division that returns ArithmeticError instead of wrapping around on overflow.

EnsureDivAssign

Performs self division that returns ArithmeticError instead of wrapping around on overflow.

EnsureFixedPointNumber

Extends FixedPointNumber with the Ensure family functions.

EnsureFrom

Similar to TryFrom but returning an ArithmeticError error.

EnsureInherentsAreFirst

A trait to ensure the inherent are before non-inherent in a block.

EnsureInto

Similar to TryInto but returning an ArithmeticError error.

EnsureMul

Performs multiplication that returns ArithmeticError instead of wrapping around on overflow.

EnsureMulAssign

Performs self multiplication that returns ArithmeticError instead of wrapping around on overflow.

EnsureOp

Meta trait that supports all immutable arithmetic Ensure* operations

EnsureOpAssign

Meta trait that supports all assigned arithmetic Ensure* operations

EnsureOrigin

Some sort of check on the origin is performed by this object.

EnsureOriginWithArg

Some sort of check on the origin is performed by this object.

EnsureSub

Performs subtraction that returns ArithmeticError instead of wrapping around on underflow.

EnsureSubAssign

Performs self subtraction that returns ArithmeticError instead of wrapping around on underflow.

EstimateCallFee

Something that can estimate the fee of a (frame-based) call.

EstimateNextNewSession

Something that can estimate at which block scheduling of the next session will happen (i.e when we will try to fetch new validators).

EstimateNextSessionRotation

Something that can estimate at which block the next session rotation will happen (i.e. a new session starts).

ExecuteBlock

Something that can execute a given block.

Extrinsic

Something that acts like an Extrinsic.

ExtrinsicCall

An extrinsic on which we can get access to call.

ExtrinsicMetadata

Implementor is an Extrinsic and provides metadata about this extrinsic.

FilterDeprecated

FilterStack

Trait to add a constraint onto the filter.

FindAuthor

A trait for finding the author of a block header based on the PreRuntime digests contained within it.

GenesisBuildDeprecated

A trait to define the build function of a genesis config, T and I are placeholder for pallet trait and pallet instance.

Get

A trait for querying a single value from a type.

GetBacking

Retrieve the backing from an object’s ref.

GetCallIndex

Gets the function index of the Call.

GetCallMetadata

Gets the metadata for the Call - function name and pallet name.

GetCallName

Gets the function name of the Call.

GetNodeBlockType

A marker trait for something that knows the type of the node block.

GetRuntimeBlockType

A marker trait for something that knows the type of the runtime block.

GetStorageVersion

Provides information about the storage version of a pallet.

HandleLifetime

A simple, generic one-parameter event notifier/handler.

HandleMessage

Handles incoming messages for a single origin.

Hash

Abstraction around hashing

HashOutput

Super trait with all the attributes for a hashing output.

Header

Something which fulfills the abstract idea of a Substrate header. It has types for a Number, a Hash and a Hashing. It provides access to an extrinsics_root, state_root and parent_hash, as well as a digest and a block number.

Hooks

The pallet hooks trait. This is merely an umbrella trait for:

IdentifyAccount

Some type that is able to be collapsed into an account ID. It is not possible to recreate the original value from the account ID.

Imbalance

A trait for a not-quite Linear Type that tracks an imbalance.

Incrementable

A trait representing an incrementable type.

InitializeMembers

Trait for type that can handle the initialization of account IDs at genesis.

Instance

An instance of a pallet in the storage.

InstanceFilter

Simple trait for providing a filter over a reference to some type, given an instance of itself.

IntegerSquareRoot

A trait implementing integer square root.

IntegrityTest

See Hooks::integrity_test.

IsMember

Determine if a MemberId is a valid member.

IsSubType

Something that can be checked to be a of sub type T.

IsType

Trait to be used when types are exactly same.

KeyOwnerProofSystem

Something which can compute and check proofs of a historical key owner and return full identification data of that key owner.

Lateness

Trait to be used by block producing consensus engine modules to determine how late the current block is (e.g. in a slot-based proposal mechanism how many slots were skipped since the previous block).

Lazy

A lazy value.

Len

Anything that can have a ::len() method.

LockableCurrency

A currency whose accounts can have liquidity restrictions.

Locker

Trait to handle NFT locking mechanism to ensure interactions with the asset can be implemented downstream to extend logic of Uniques/Nfts current functionality.

Lookup

Means of changing one type into another in a manner dependent on the source type.

MaybeConvert

Fallible conversion trait returning an Option. Generic over both source and destination types.

MaybeConvertBack

Reversing fallible conversion trait returning an Option. Generic over both source and destination types.

MaybeDisplay

A type that implements Display when in std environment.

MaybeEquivalence

Definition for a bi-directional, fallible conversion between two types.

MaybeFromStr

A type that implements FromStr when in std environment.

MaybeHash

A type that implements Hash when in std environment.

MaybeSerialize

A type that implements Serialize when in std environment or serde feature is activated.

MaybeSerializeDeserialize

A type that implements Serialize, DeserializeOwned and Debug when in std environment or serde feature is activated.

Member

A type that can be used in runtime structures.

Morph

Extensible conversion trait. Generic over only source type, with destination type being associated.

NamedReservableCurrency

OffchainWorker

Off-chain computation trait.

OnFinalize

See Hooks::on_finalize.

OnGenesis

A trait that will be called at genesis.

OnIdle

See Hooks::on_idle.

OnInitialize

See Hooks::on_initialize.

OnKilledAccount

The account with the given id was reaped.

OnNewAccount

Handler for when a new account has been created.

OnRuntimeUpgrade

See Hooks::on_runtime_upgrade.

OnTimestampSet

A trait which is called when the timestamp is set in the runtime.

OnUnbalanced

Handler for when some currency “account” decreased in balance for some reason.

One

Defines a multiplicative identity element for Self.

OneSessionHandler

A session handler for specific key type.

OpaqueKeys

Opaque data type that may be destructured into a series of raw byte slices (which represent individual keys).

OriginTrait

Methods available on frame_system::Config::RuntimeOrigin.

PalletError

Trait indicating that the implementing type is going to be included as a field in a variant of the #[pallet::error] enum type.

PalletInfo

Provides information about the pallet itself and its setup in the runtime.

PalletInfoAccess

Provides information about the pallet itself and its setup in the runtime.

PalletsInfoAccess

Provide information about a bunch of pallets.

PartialStorageInfoTrait

Similar to StorageInfoTrait, a trait to give partial information about storage.

Polling

PreimageProvider

A interface for looking up preimages from their hash on chain.

PreimageRecipient

A interface for managing preimages to hashes on chain.

Printable

Trait for things which can be printed from the runtime.

PrivilegeCmp

Something that can compare privileges of two origins.

ProcessMessage

Can process messages from a specific origin.

QueryPreimage

A interface for looking up preimages from their hash on chain.

QueuePausedQuery

Provides information on paused queues.

Randomness

A trait that is able to provide randomness.

RankedMembers

Ranked membership data structure.

ReservableCurrency

A currency where funds can be reserved from the user.

SafeMode

Can put the runtime into a safe mode.

SafeModeNotify

A trait to notify when the runtime enters or exits safe mode.

SaturatedConversion

Convenience type to work around the highly unergonomic syntax needed to invoke the functions of overloaded generic traits, in this case SaturatedFrom and SaturatedInto.

Saturating

Saturating arithmetic operations, returning maximum or minimum values instead of overflowing.

Scale

Multiply and divide by a number that isn’t necessarily the same type. Basically just the same as Mul and Div except it can be used for all basic numeric types.

ServiceQueues

Can service queues and execute overweight messages.

SignaturePayload

Something that acts like a SignaturePayload of an Extrinsic.

SignedExtension

Means by which a transaction may be extended. This type embodies both the data and the logic that should be additionally associated with the transaction. It should be plain old data.

SimpleBitOps

A meta trait for all bit ops.

SortedMembers

A trait for a set which can enumerate its members in order.

StaticLookup

Means of changing one type into another in a manner dependent on the source type. This variant is different to Lookup in that it doesn’t (can cannot) require any context.

StorageInfoTrait

A trait to give information about storage.

StorageInstance

An instance of a storage in a pallet.

StorePreimage

A interface for managing preimages to hashes on chain.

StoredMap

An abstraction of a value stored within storage, but possibly as part of a larger composite item.

Time

TransactionPause

Can pause specific transactions from being processed.

TryCollect

Try and collect into a collection C.

TryConvert

Fallible conversion trait which returns the argument in the case of being unable to convert. Generic over both source and destination types.

TryConvertBack

Reversing fallible conversion trait which returns the argument in the case of being unable to convert back. Generic over both source and destination types.

TryDrop

A type for which some values make sense to be able to drop without further consideration.

TryMorph

Extensible conversion trait. Generic over only source type, with destination type being associated.

TypedGet

A trait for querying a single value from a type defined in the trait.

UnfilteredDispatchable

Type that can be dispatched with an origin but without checking the origin filter.

UniqueSaturatedFrom

Just like From except that if the source value is too big to fit into the destination type then it’ll saturate the destination.

UniqueSaturatedInto

Just like Into except that if the source value is too big to fit into the destination type then it’ll saturate the destination.

UnixTime

Trait to deal with unix time.

ValidateUnsigned

Provide validation for unsigned extrinsics.

ValidatorRegistration

Implementors of this trait provide information about whether or not some validator has been registered with them. The Session module is an implementor.

ValidatorSet

A trait for online node inspection in a session.

ValidatorSetWithIdentification

ValidatorSet combined with an identification.

VariantCount

Trait to get the number of variants in any enum.

Verify

Means of signature verification.

VerifySeal

A trait for verifying the seal of a header and returning the author.

VestingSchedule

A vesting schedule over a currency. This allows a particular currency to have vesting limits applied to it.

VoteTally

WhitelistedStorageKeys

Allows a pallet to specify storage keys to whitelist during benchmarking. This means those keys will be excluded from the benchmarking performance calculation.

Zero

Defines an additive identity element for Self.

Functions

checked_pow

Raises a value to the power of exp, returning None if an overflow occurred.

ensure_pow

Raises a value to the power of exp, returning ArithmeticError if an overflow occurred.

Type Definitions

AllowAllDeprecated

BoundedInline

DenyAllDeprecated

DispatchInfoOf

Shortcut to reference the Info type of a Dispatchable.

EnsureOneOfDeprecated

“OR gate” implementation of EnsureOrigin allowing for different Success types for L and R, with them combined using an Either type.

FetchResult

HashingFor

Extract the hashing type for a block.

LockIdentifier

An identifier for a lock. Used for disambiguating different locks so that they can be individually replaced or removed.

NumberFor

Extract the number type for a block.

PostDispatchInfoOf

Shortcut to reference the PostInfo type of a Dispatchable.