Module frame::traits::tokens::fungible

The traits for dealing with a single fungible token class and any associated types.

User-implememted traits

Inspect: Regular balance inspector functions.

Unbalanced: Low-level balance mutating functions. Does not guarantee proper book-keeping and so should not be called into directly from application code. Other traits depend on this and provide default implementations based on it.

UnbalancedHold: Low-level balance mutating functions for balances placed on hold. Does not guarantee proper book-keeping and so should not be called into directly from application code. Other traits depend on this and provide default implementations based on it.

Mutate: Regular balance mutator functions. Pre-implemented using Unbalanced, though the done_* functions should likely be reimplemented in case you want to do something following the operation such as emit events.

InspectHold: Inspector functions for balances on hold.

MutateHold: Mutator functions for balances on hold. Mostly pre-implemented using UnbalancedHold.

InspectFreeze: Inspector functions for frozen balance.

MutateFreeze: Mutator functions for frozen balance.

Balanced: One-sided mutator functions for regular balances, which return imbalance objects which guarantee eventual book-keeping. May be useful for some sophisticated operations where funds must be removed from an account before it is known precisely what should be done with them.

Modules

conformance_tests

freeze

The traits for putting freezes within a single fungible token class.

hold

The traits for putting holds within a single fungible token class.

Structs

DecreaseIssuance

Simple handler for an imbalance drop which decreases the total issuance of the system by the imbalance amount. Used for leftover credit.

Dust

Special dust type which can be type-safely converted into a Credit.

FreezeConsideration

Consideration method using a fungible balance frozen as the cost exacted for the footprint.

HoldConsideration

Consideration method using a fungible balance frozen as the cost exacted for the footprint.

Imbalance

An imbalance in the system, representing a divergence of recorded token supply from the sum of the balances of all accounts. This is must_use in order to ensure it gets handled (placing into an account, settling from an account or altering the supply).

IncreaseIssuance

Simple handler for an imbalance drop which increases the total issuance of the system by the imbalance amount. Used for leftover debt.

ItemOf

Convert a fungibles trait implementation into a fungible trait implementation by identifying a single item.

LoneFreezeConsideration

Basic consideration method using a fungible balance frozen as the cost exacted for the footprint.

LoneHoldConsideration

Basic consideration method using a fungible balance placed on hold as the cost exacted for the footprint.

Traits

Balanced

A fungible token class where any creation and deletion of tokens is semi-explicit and where the total supply is maintained automatically.

BalancedHold

Trait for slashing a fungible asset which can be place on hold.

HandleImbalanceDrop

Handler for when an imbalance gets dropped. This could handle either a credit (negative) or debt (positive) imbalance.

Inspect

Trait for providing balance-inspection access to a fungible asset.

InspectFreeze

Trait for inspecting a fungible asset which can be frozen. Freezing is essentially setting a minimum balance bellow which the total balance (inclusive of any funds placed on hold) may not be normally allowed to drop. Generally, freezers will provide an “update” function such that if the total balance does drop below the limit, then the freezer can update their housekeeping accordingly.

InspectHold

Trait for inspecting a fungible asset whose accounts support partitioning and slashing.

Mutate

Trait for providing a basic fungible asset.

MutateFreeze

Trait for introducing, altering and removing locks to freeze an account’s funds so they never go below a set minimum.

MutateHold

Trait for mutating a fungible asset which can be placed on hold.

Unbalanced

A fungible token class where the balance can be set arbitrarily.

UnbalancedHold

A fungible, holdable token class where the balance on hold can be set arbitrarily.

Type Definitions

Credit

Imbalance implying that the total_issuance value is greater than the sum of all account balances.

Debt

Imbalance implying that the total_issuance value is less than the sum of all account balances.