Struct frame::traits::tokens::fungible::Imbalance

pub struct Imbalance<B, OnDrop, OppositeOnDrop>where
    B: Balance,
    OnDrop: HandleImbalanceDrop<B>,
    OppositeOnDrop: HandleImbalanceDrop<B>,{ /* private fields */ }

An imbalance in the system, representing a divergence of recorded token supply from the sum of the balances of all accounts. This is must_use in order to ensure it gets handled (placing into an account, settling from an account or altering the supply).

Importantly, it has a special Drop impl, and cannot be created outside of this module.