Type Alias frame::traits::tokens::fungibles::Debt

pub type Debt<AccountId, B> = Imbalance<<B as Inspect<AccountId>>::AssetId, <B as Inspect<AccountId>>::Balance, <B as Balanced<AccountId>>::OnDropDebt, <B as Balanced<AccountId>>::OnDropCredit>;

Imbalance implying that the total_issuance value is less than the sum of all account balances.

Aliased Type

struct Debt<AccountId, B> { /* private fields */ }

Implementations

impl<A, B, OnDrop, OppositeOnDrop> Imbalance<A, B, OnDrop, OppositeOnDrop>where A: AssetId, B: Balance, OnDrop: HandleImbalanceDrop<A, B>, OppositeOnDrop: HandleImbalanceDrop<A, B>,

pub fn zero(asset: A) -> Imbalance<A, B, OnDrop, OppositeOnDrop>

pub fn drop_zero(self) -> Result<(), Imbalance<A, B, OnDrop, OppositeOnDrop>>

pub fn split( self, amount: B ) -> (Imbalance<A, B, OnDrop, OppositeOnDrop>, Imbalance<A, B, OnDrop, OppositeOnDrop>)

pub fn extract(&mut self, amount: B) -> Imbalance<A, B, OnDrop, OppositeOnDrop>

Mutate self by extracting a new instance with at most amount value, reducing self accordingly.

pub fn merge( self, other: Imbalance<A, B, OnDrop, OppositeOnDrop> ) -> Result<Imbalance<A, B, OnDrop, OppositeOnDrop>, (Imbalance<A, B, OnDrop, OppositeOnDrop>, Imbalance<A, B, OnDrop, OppositeOnDrop>)>

pub fn subsume( &mut self, other: Imbalance<A, B, OnDrop, OppositeOnDrop> ) -> Result<(), Imbalance<A, B, OnDrop, OppositeOnDrop>>

pub fn offset( self, other: Imbalance<A, B, OppositeOnDrop, OnDrop> ) -> Result<SameOrOther<Imbalance<A, B, OnDrop, OppositeOnDrop>, Imbalance<A, B, OppositeOnDrop, OnDrop>>, (Imbalance<A, B, OnDrop, OppositeOnDrop>, Imbalance<A, B, OppositeOnDrop, OnDrop>)>

pub fn peek(&self) -> B

pub fn asset(&self) -> A