Trait frame::traits::EnsureInherentsAreFirst

pub trait EnsureInherentsAreFirst<Block> {
    // Required method
    fn ensure_inherents_are_first(block: &Block) -> Result<(), u32>;
}

A trait to ensure the inherent are before non-inherent in a block.

This is typically implemented on runtime, through construct_runtime!.

Required Methods

fn ensure_inherents_are_first(block: &Block) -> Result<(), u32>

Ensure the position of inherent is correct, i.e. they are before non-inherents.

On error return the index of the inherent with invalid position (counting from 0).