Trait frame::traits::EnsureOriginWithArg

pub trait EnsureOriginWithArg<OuterOrigin, Argument> {
    type Success;

    // Required method
    fn try_origin(
        o: OuterOrigin,
        a: &Argument
    ) -> Result<Self::Success, OuterOrigin>;

    // Provided method
    fn ensure_origin(
        o: OuterOrigin,
        a: &Argument
    ) -> Result<Self::Success, BadOrigin> { ... }
}

Some sort of check on the origin is performed by this object.

Required Associated Types

type Success

A return type.

Required Methods

fn try_origin( o: OuterOrigin, a: &Argument ) -> Result<Self::Success, OuterOrigin>

Perform the origin check, returning the origin value if unsuccessful. This allows chaining.

Provided Methods

fn ensure_origin( o: OuterOrigin, a: &Argument ) -> Result<Self::Success, BadOrigin>

Perform the origin check.