Trait frame::traits::ExecuteBlock

pub trait ExecuteBlock<Block>where
    Block: Block,{
    // Required method
    fn execute_block(block: Block);
}

Something that can execute a given block.

Executing a block means that all extrinsics in a given block will be executed and the resulting header will be checked against the header of the given block.

Required Methods

fn execute_block(block: Block)

Execute the given block.

This will execute all extrinsics in the block and check that the resulting header is correct.

Panic

Panics when an extrinsics panics or the resulting header doesn’t match the expected header.