Trait frame::traits::Incrementable

pub trait Incrementable: Sized {
    // Required methods
    fn increment(&self) -> Option<Self>;
    fn initial_value() -> Option<Self>;
}

A trait representing an incrementable type.

The increment and initial_value functions are fallible. They should either both return Some with a valid value, or None.

Required Methods

fn increment(&self) -> Option<Self>

Increments the value.

Returns Some with the incremented value if it is possible, or None if it is not.

fn initial_value() -> Option<Self>

Returns the initial value.

Returns Some with the initial value if it is available, or None if it is not.