Trait frame::traits::PreInherents

pub trait PreInherents {
    // Provided method
    fn pre_inherents() { ... }
}

Provides a callback to execute logic before the all inherents.

Provided Methods

fn pre_inherents()

Called before all inherents were applied but after on_initialize.

Object Safety

This trait is not object safe.