Trait frame::traits::TransactionExtension

pub trait TransactionExtension<Call, Context>: TransactionExtensionBasewhere
    Call: Dispatchable,{
    type Val;
    type Pre;

    // Required methods
    fn validate(
        &self,
        origin: <Call as Dispatchable>::RuntimeOrigin,
        call: &Call,
        info: &<Call as Dispatchable>::Info,
        len: usize,
        context: &mut Context,
        self_implicit: Self::Implicit,
        inherited_implication: &impl Encode
    ) -> Result<(ValidTransaction, Self::Val, <Call as Dispatchable>::RuntimeOrigin), TransactionValidityError>;
    fn prepare(
        self,
        val: Self::Val,
        origin: &<Call as Dispatchable>::RuntimeOrigin,
        call: &Call,
        info: &<Call as Dispatchable>::Info,
        len: usize,
        context: &Context
    ) -> Result<Self::Pre, TransactionValidityError>;

    // Provided methods
    fn post_dispatch(
        _pre: Self::Pre,
        _info: &<Call as Dispatchable>::Info,
        _post_info: &<Call as Dispatchable>::PostInfo,
        _len: usize,
        _result: &Result<(), DispatchError>,
        _context: &Context
    ) -> Result<(), TransactionValidityError> { ... }
    fn validate_bare_compat(
        _call: &Call,
        _info: &<Call as Dispatchable>::Info,
        _len: usize
    ) -> Result<ValidTransaction, TransactionValidityError> { ... }
    fn pre_dispatch_bare_compat(
        _call: &Call,
        _info: &<Call as Dispatchable>::Info,
        _len: usize
    ) -> Result<(), TransactionValidityError> { ... }
    fn post_dispatch_bare_compat(
        _info: &<Call as Dispatchable>::Info,
        _post_info: &<Call as Dispatchable>::PostInfo,
        _len: usize,
        _result: &Result<(), DispatchError>
    ) -> Result<(), TransactionValidityError> { ... }
}

Means by which a transaction may be extended. This type embodies both the data and the logic that should be additionally associated with the transaction. It should be plain old data.

The simplest transaction extension would be the Unit type (and empty pipeline) (). This executes no additional logic and implies a dispatch of the transaction’s call using the inherited origin (either None or Signed, depending on whether this is a signed or general transaction).

Transaction extensions are capable of altering certain associated semantics:

They may define the origin with which the transaction’s call should be dispatched.

They may define various parameters used by the transction queue to determine under what conditions the transaction should be retained and introduced on-chain.

They may define whether this transaction is acceptable for introduction on-chain at all.

Each of these semantics are defined by the validate function.

NOTE: Transaction extensions cannot under any circumctances alter the call itself.

Transaction extensions are capable of defining logic which is executed additionally to the dispatch of the call:

They may define logic which must be executed prior to the dispatch of the call.

They may also define logic which must be executed after the dispatch of the call.

Each of these semantics are defined by the prepare and post_dispatch functions respectively.

Finally, transaction extensions may define additional data to help define the implications of the logic they introduce. This additional data may be explicitly defined by the transaction author (in which case it is included as part of the transaction body), or it may be implicitly defined by the transaction extension based around the on-chain state (which the transaction author is assumed to know). This data may be utilized by the above logic to alter how a node’s transaction queue treats this transaction.

Default implementations

Of the 5 functions in this trait, 3 of them must return a value of an associated type on success, and none of these types implement Default or anything like it. This means that default implementations cannot be provided for these functions. However, a macro is provided impl_tx_ext_default which is capable of generating default implementations for each of these 3 functions. If you do not wish to introduce additional logic into the transaction pipeline, then it is recommended that you use this macro to implement these functions.

Pipelines, Inherited Implications, and Authorized Origins

Requiring a single transaction extension to define all of the above semantics would be cumbersome and would lead to a lot of boilerplate. Instead, transaction extensions are aggregated into pipelines, which are tuples of transaction extensions. Each extension in the pipeline is executed in order, and the output of each extension is aggregated and/or relayed as the input to the next extension in the pipeline.

This ordered composition happens with all datatypes (Val, Pre and Implicit) as well as all functions. There are important consequences stemming from how the composition affects the meaning of the origin and implication parameters as well as the results. Whereas the prepare and post_dispatch functions are clear in their meaning, the validate function is fairly sophisticated and warrants further explanation.

Firstly, the origin parameter. The origin passed into the first item in a pipeline is simply that passed into the tuple itself. It represents an authority who has authorized the implication of the transaction, as of the extension it has been passed into *and any further extensions it may pass though, all the way to, and including, the transaction’s dispatch call itself. Each following item in the pipeline is passed the origin which the previous item returned. The origin returned from the final item in the pipeline is the origin which is returned by the tuple itself.

This means that if a constituent extension returns a different origin to the one it was called with, then (assuming no other extension changes it further) this new origin will be used for all extensions following it in the pipeline, and will be returned from the pipeline to be used as the origin for the call’s dispatch. The call itself as well as all these extensions following may each imply consequence for this origin. We call this the inherited implication.

The inherited implication is the cumulated on-chain effects born by whatever origin is returned. It is expressed to the validate function only as the implication argument which implements the Encode trait. A transaction extension may define its own implications through its own fields and the implicit function. This is only utilized by extensions which preceed it in a pipeline or, if the transaction is an old-school signed trasnaction, the underlying transaction verification logic.

The inherited implication passed as the implication parameter to validate does not include the extension’s inner data itself nor does it include the result of the extension’s implicit function. If you both provide an implication and rely on the implication, then you need to manually aggregate your extensions implication with the aggregated implication passed in.

Required Associated Types

type Val

The type that encodes information that can be passed from validate to prepare.

type Pre

The type that encodes information that can be passed from prepare to post-dispatch.

Required Methods

fn validate( &self, origin: <Call as Dispatchable>::RuntimeOrigin, call: &Call, info: &<Call as Dispatchable>::Info, len: usize, context: &mut Context, self_implicit: Self::Implicit, inherited_implication: &impl Encode ) -> Result<(ValidTransaction, Self::Val, <Call as Dispatchable>::RuntimeOrigin), TransactionValidityError>

Validate a transaction for the transaction queue.

This function can be called frequently by the transaction queue to obtain transaction validity against current state. It should perform all checks that determine a valid transaction, that can pay for its execution and quickly eliminate ones that are stale or incorrect.

Parameters:

origin: The origin of the transaction which this extension inherited; coming from an “old-school” signed transaction, this will be a system RawOrigin::Signed value. If the transaction is a “new-school” General Transaction, then this will be a system RawOrigin::None value. If this extension is an item in a composite, then it could be anything which was previously returned as an origin value in the result of a validate call.

call: The Call wrapped by this extension.

info: Information concerning, and inherent to, the transaction’s call.

len: The total length of the encoded transaction.

inherited_implication: The implication which this extension inherits. This is a tuple of the transaction’s call and some additional opaque-but-encodable data. Coming directly from a transaction, the latter is [()]. However, if this extension is expressed as part of a composite type, then the latter component is equal to any further implications to which the returned origin could potentially apply. See Pipelines, Inherited Implications, and Authorized Origins for more information.

context: Some opaque mutable context, as yet unused.

Returns a ValidateResult, which is a Result whose success type is a tuple of ValidTransaction (defining useful metadata for the transaction queue), the Self::Val token of this transaction, which gets passed into prepare, and the origin of the transaction, which gets passed into prepare and is ultimately used for dispatch.

fn prepare( self, val: Self::Val, origin: &<Call as Dispatchable>::RuntimeOrigin, call: &Call, info: &<Call as Dispatchable>::Info, len: usize, context: &Context ) -> Result<Self::Pre, TransactionValidityError>

Do any pre-flight stuff for a transaction after validation.

This is for actions which do not happen in the transaction queue but only immediately prior to the point of dispatch on-chain. This should not return an error, since errors should already have been identified during the validate call. If an error is returned, the transaction will be considered invalid but no state changes will happen and therefore work done in validate will not be paid for.

Unlike validate, this function may consume self.

Parameters:

val: Self::Val returned by the result of the validate call.

origin: The origin returned by the result of the validate call.

call: The Call wrapped by this extension.

info: Information concerning, and inherent to, the transaction’s call.

len: The total length of the encoded transaction.

context: Some opaque mutable context, as yet unused.

Returns a Self::Pre value on success, which gets passed into post_dispatch and after the call is dispatched.

IMPORTANT: Checks made in validation need not be repeated here.

Provided Methods

fn post_dispatch( _pre: Self::Pre, _info: &<Call as Dispatchable>::Info, _post_info: &<Call as Dispatchable>::PostInfo, _len: usize, _result: &Result<(), DispatchError>, _context: &Context ) -> Result<(), TransactionValidityError>

Do any post-flight stuff for an extrinsic.

_pre contains the output of prepare.

This gets given the DispatchResult _result from the extrinsic and can, if desired, introduce a TransactionValidityError, causing the block to become invalid for including it.

Parameters:

pre: Self::Pre returned by the result of the prepare call prior to dispatch.

info: Information concerning, and inherent to, the transaction’s call.

post_info: Information concerning the dispatch of the transaction’s call.

len: The total length of the encoded transaction.

result: The result of the dispatch.

context: Some opaque mutable context, as yet unused.

WARNING: It is dangerous to return an error here. To do so will fundamentally invalidate the transaction and any block that it is included in, causing the block author to not be compensated for their work in validating the transaction or producing the block so far. It can only be used safely when you know that the transaction is one that would only be introduced by the current block author.

fn validate_bare_compat( _call: &Call, _info: &<Call as Dispatchable>::Info, _len: usize ) -> Result<ValidTransaction, TransactionValidityError>

👎Deprecated: Only for compatibility. DO NOT USE.

Compatibility function for supporting the SignedExtension::validate_unsigned function.

DO NOT USE! THIS MAY BE REMOVED AT ANY TIME!

fn pre_dispatch_bare_compat( _call: &Call, _info: &<Call as Dispatchable>::Info, _len: usize ) -> Result<(), TransactionValidityError>

👎Deprecated: Only for compatibility. DO NOT USE.

Compatibility function for supporting the SignedExtension::pre_dispatch_unsigned function.

DO NOT USE! THIS MAY BE REMOVED AT ANY TIME!

fn post_dispatch_bare_compat( _info: &<Call as Dispatchable>::Info, _post_info: &<Call as Dispatchable>::PostInfo, _len: usize, _result: &Result<(), DispatchError> ) -> Result<(), TransactionValidityError>

👎Deprecated: Only for compatibility. DO NOT USE.

Compatibility function for supporting the SignedExtension::post_dispatch function where pre is None.

DO NOT USE! THIS MAY BE REMOVED AT ANY TIME!

Object Safety

This trait is not object safe.