Trait frame::traits::transaction_extension::TransactionExtensionBase

pub trait TransactionExtensionBase: TransactionExtensionInterior {
    type Implicit: Codec + StaticTypeInfo;

    const IDENTIFIER: &'static str;

    // Provided methods
    fn implicit(&self) -> Result<Self::Implicit, TransactionValidityError> { ... }
    fn weight(&self) -> Weight { ... }
    fn metadata() -> Vec<TransactionExtensionMetadata> { ... }
}

Base for TransactionExtensions; this contains the associated types and does not require any generic parameterization.

Required Associated Types

type Implicit: Codec + StaticTypeInfo

Any additional data which was known at the time of transaction construction and can be useful in authenticating the transaction. This is determined dynamically in part from the on-chain environment using the implicit function and not directly contained in the transaction itself and therefore is considered “implicit”.

Required Associated Constants

const IDENTIFIER: &'static str

Unique identifier of this signed extension.

This will be exposed in the metadata to identify the signed extension used in an extrinsic.

Provided Methods

fn implicit(&self) -> Result<Self::Implicit, TransactionValidityError>

Determine any additional data which was known at the time of transaction construction and can be useful in authenticating the transaction. The expected usage of this is to include in any data which is signed and verified as part of transactiob validation. Also perform any pre-signature-verification checks and return an error if needed.

fn weight(&self) -> Weight

The weight consumed by executing this extension instance fully during transaction dispatch.

fn metadata() -> Vec<TransactionExtensionMetadata>

Returns the metadata for this extension.

As a TransactionExtension can be a tuple of TransactionExtensions we need to return a Vec that holds the metadata of each one. Each individual TransactionExtension must return exactly one TransactionExtensionMetadata.

This method provides a default implementation that returns a vec containing a single TransactionExtensionMetadata.

Object Safety

This trait is not object safe.