Crate frame

FRAME

  ______   ______    ________   ___ __ __   ______
 /_____/\ /_____/\  /_______/\ /__//_//_/\ /_____/\
 \::::_\/_\:::_ \ \ \::: _  \ \\::\| \| \ \\::::_\/_
  \:\/___/\\:(_) ) )_\::(_)  \ \\:.      \ \\:\/___/\
   \:::._\/ \: __ `\ \\:: __  \ \\:.\-/\  \ \\::___\/_
    \:\ \    \ \ `\ \ \\:.\ \  \ \\. \  \  \ \\:\____/\
     \_\/     \_\/ \_\/ \__\/\__\/ \__\/ \__\/ \_____\/

Framework for Runtime Aggregation of Modularized Entities: Substrate’s State Transition Function (Runtime) Framework.

Documentation

TODO: Link to devhub

Warning: Experimental

This crate and all of its content is experimental, and should not yet be used in production.

Underlying dependencies

This crate is an amalgamation of multiple other crates that are often used together to compose a pallet. It is not necessary to use it, and it may fall short for certain purposes.

In short, this crate only re-exports types and traits from multiple sources. All of these sources are listed (and re-exported again) in deps.

Re-exports

pub use log;

Modules

arithmetic

The arithmetic types used for safe math.

deps

Access to all of the dependencies of this crate. In case the re-exports are not enough, this module can be used.

derive

All derive macros used in frame.

pallet_macros

A list of all macros used within the main pallet macro.

prelude

The main prelude of FRAME.

primitives

Low level primitive types used in FRAME pallets.

runtime

All of the types and tools needed to build FRAME-based runtimes.

testing_prelude

The main testing prelude of FRAME.

traits

All traits often used in FRAME pallets.

Attribute Macros

import_section

An attribute macro that can be attached to a module declaration. Doing so will Imports the contents of the specified external pallet section that was defined previously using #[pallet_section].

pallet

Exports the main pallet macro. This can wrap a mod pallet and will transform it into being a pallet, eg #[frame::pallet] mod pallet { .. }.

pallet_section

Can be attached to a module. Doing so will declare that module as importable into a pallet via #[import_section].