Trait frame::traits::UnfilteredDispatchable

pub trait UnfilteredDispatchable {
    type RuntimeOrigin;

    // Required method
    fn dispatch_bypass_filter(
        self,
        origin: Self::RuntimeOrigin
    ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>;
}

Type that can be dispatched with an origin but without checking the origin filter.

Implemented for pallet dispatchable type by decl_module and for runtime dispatchable by construct_runtime.

Required Associated Types

type RuntimeOrigin

The origin type of the runtime, (i.e. frame_system::Config::RuntimeOrigin).

Required Methods

fn dispatch_bypass_filter( self, origin: Self::RuntimeOrigin ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>

Dispatch this call but do not check the filter in origin.